[Persistence] Utilisation de Hibernate avec QueryDsl pour les requêtes dans CompanyService.

+++ src/main/java/com/excilys/cdb/persistence/ComputerDao.java
@@ -49,5 +49,5 @@ public interface ComputerDao extends CommonDao<Computer> {
 	 * @param id
 	 *            Company identifier.
 	 */
-	public void deleteComputersByCompanyId(int companyId);
+	public void deleteComputersByCompanyId(long companyId);
 }
